requirements_airflow.txt
tqdm
pydantic
pandas
python-dotenv
google-cloud-storage
apache-airflow==3.0.1
apache-airflow-providers-apache-spark
pyspark==3.5.5
psycopg2-binary
asyncpg
apache-airflow-providers-fab
apache-airflow-providers-celery